Adding ROS2 support

+++ include/ImuTypes.h
@@ -222,7 +222,7 @@ public:
 
     void printMeasurements() const {
         std::cout << "pint meas:\n";
-        for(int i=0; i<mvMeasurements.size(); i++){
+        for(size_t i=0; i<mvMeasurements.size(); i++){
             std::cout << "meas " << mvMeasurements[i].t << std::endl;
         }
         std::cout << "end pint meas:\n";
